Move remote middleware tests out
